update button classes to reduce file size & increase reusability
In an effort to cut down on file size & make the button classes more modular, I reduced the selectors of the buttons to no longer be attached to the <button> element.

now .btn-a, b, c, & small can be combined with a generic .btn class (which has the same styles as <button> and this class combo can be added to <a> elements, inputs with a type of submit or anything that one would like to look like a button.

I used terminal to run a gzip compression on the minified file and it also reduced the file size by 2 bytes